watermark_service/requirements.txt
Pillow==11.1.0
numpy==1.26.4
blind-watermark==0.4.4
opencv-python==4.10.0.84
